feat(treasury): add Markdown preview tab in create proposal page (#267)
* feat(treasury): add Markdown preview tab in create proposal page

This commit adds a Markdown preview tab to the `create-proposal` page. This allows
users to view their proposal descriptions in Markdown format in real-time, improving
the user experience by providing immediate visual feedback.

* refactor(treasury): improve markdown preview styling

This commit updates the treasury proposal markdown styling so that it is
more in line with what is shown when the proposal is live.

* feat: add title to treasury markdown preview

This commit adds the title to the treasury markdown preview found on the
create-proposal page.